backend_smart_presence/requirements.txt
# === Core API ===
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
requests==2.31.0

# === Database (Supabase / Postgres) ===
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0

# === Authentication & Security ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
email-validator==2.1.0.post1

# === AI & Computer Vision ===
# Runtime package is auto-switched by setup_env.bat:
# - CPU: onnxruntime
# - GPU: onnxruntime-gpu (when NVIDIA/CUDA is detected)
numpy<2.0.0
opencv-python-headless==4.8.1.78
onnxruntime==1.16.3
insightface==0.7.3

# === Vector Storage (Local) ===
chromadb==0.4.22

# === Testing ===
pytest==7.4.3
httpx==0.25.2
